refactor: MemoryRoomRepository totalCount 반환 타입 변경
- totalCount 메서드의 반환 타입을 int에서 long으로 변경
- 대규모 데이터를 다룰 때의 정확성 보장

+++ backend/src/main/java/coffeeshout/room/domain/repository/MemoryRoomRepository.java
@@ -54,7 +54,7 @@ public class MemoryRoomRepository implements RoomRepository {
     /**
      * 전체 Room 수를 반환한다.
      */
-    public int totalCount() {
+    public long totalCount() {
         return rooms.size();
     }
 }
